feat: enum 파일 재적용
