+++ nest/src/common/helpers/enum.helper.ts
@@ -16,11 +16,3 @@ export enum ServerEnviroment {
   TEST = 'test',
   PROD = 'production',
 }
-
-export enum ChatEvent {
-  ROOM_ENTER = 'enter',
-  ROOM_EXIT = 'exit',
-  CHECK_MULTIPLE_USER = 'check_multiple_user',
-  SEND_CHAT = 'chat',
-  SEND_MEMBER_COUNT = 'member_count',
-}
