+++ nest/src/models/chat/interface/enum.ts
@@ -1,0 +1,8 @@
+export enum ChatEvent {
+  ROOM_ENTER = 'enter',
+  ROOM_EXIT = 'exit',
+  CHECK_MULTIPLE_USER = 'check_multiple_user',
+  SEND_CHAT = 'chat',
+  SEND_MEMBER_COUNT = 'member_count',
+  WELCOME_MESSAGE = 'welcome_message',
+}
